Made a recovery

+++ Spice/Models/MenuItem.cs
@@ -32,7 +32,7 @@ namespace Spice.Models
         [ForeignKey("SubCategoryId")]
         public virtual SubCategory SubCategory { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage =" Price should be greater than ${1} ")]
+        [Range(1, double.MaxValue, ErrorMessage =" Price should be greater than ${1} ")]
         public double Price { get; set; }
 
     }
